Add XDamageSubtractAndTrigger operation
XDamageSubtractAndTrigger behaves exactly like
XDamageSubtract except it receives an optional
fence sync object.  If the value of this object
is not None, it is triggered by X once all the
rendering associated with the damage regions
being subtracted has completed.

Use version 1.3.0 to mark the inclusion of
DamageSubtractAndTrigger.  Really this should be
1.2.0, but that proto package version was already
released, and it's less confusing if the protocol
version and package version are in sync.

Reviewed-by: Pierre-Loup Griffais <pgriffais@nvidia.com>
Reviewed-by: Aaron Plattner <aplattner@nvidia.com>
Reviewed-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Keith Packard <keithp@keithp.com>